+++ Tests/LinuxMain.swift
@@ -8,23 +8,23 @@ XCTMain([
     testCase(ConfigTests.allTests),
     testCase(ControllerTests.allTests),
     testCase(EnvironmentTests.allTests),
-	testCase(EventTests.allTests),
+    testCase(EventTests.allTests),
     testCase(HashTests.allTests),
-	testCase(HTTPStreamTests.allTests),
+    testCase(HTTPStreamTests.allTests),
+    testCase(LocalizationTests.allTests)
     testCase(LogTests.allTests),
     testCase(MemorySessionDriverTests.allTests),
-	testCase(PerformanceTests.allTests),
+    testCase(PerformanceTests.allTests),
     testCase(ProcessTests.allTests),
     testCase(ResponseTests.allTests),
     testCase(RouterTests.allTests),
     testCase(RouteTests.allTests),
     testCase(SessionTests.allTests),
-	testCase(TypedRouteTests.allTests),
-	testCase(ValidationConvenienceTests.allTests),
-	testCase(ValidationCountTests.allTests),
-	testCase(ValidationTests.allTests),
-	testCase(ValidationUniqueTests.allTests),
-    testCase(LocalizationTests.allTests)
+    testCase(TypedRouteTests.allTests),
+    testCase(ValidationConvenienceTests.allTests),
+    testCase(ValidationCountTests.allTests),
+    testCase(ValidationTests.allTests),
+    testCase(ValidationUniqueTests.allTests),
 ])
 
 #endif
